Add files via upload
This example files PR adds the needed "internals" used for our Grafana/Prometheus setup. Currently, this is only being added for a simple cluster and a production cluster deployment but this can easily be added to other types of deployments. Adding it to the examples gives customers an option to easier roll it out.

Added functionality:

```
metrics:
    mode: sidecar
    enabled: true
    image: arangodb/arangodb-exporter:0.1.7
    tls: false
  annotations:                                                                                                                                                  
    prometheus.io/scrape: 'true'                                                                                                                               
    prometheus.io/port: '9101'                                                                                                                                 
    prometheus.io/scrape_interval: '5s'
```